Add files via upload
A generic and efficient wave spawner script for a game with multiple enemy types. Enemies will be spawned based on which enemy type is selected in the editor.. New wave will be initialized when all enemies in current wave are dead. Level will be completed when the level's wave count reaches Zero.